+++ apps/api/src/app/organization-recurring-expense/commands/organization-recurring-expense.edit.command.ts
@@ -1,4 +1,4 @@
-import { OrganizationRecurringExpenseEditInput as IExpenseEditInput } from '@gauzy/models';
+import { RecurringExpenseEditInput as IExpenseEditInput } from '@gauzy/models';
 import { ICommand } from '@nestjs/cqrs';
 
 export class OrganizationRecurringExpenseEditCommand implements ICommand {
