+++ apps/api/src/app/shared/handlers/index.ts
@@ -1,2 +1,4 @@
 export * from './update-entity.by-member.handler';
 export * from './find-recurring-expense.by-month.handler';
+export * from './recurring-expense.delete.handler';
+export * from './recurring-expense.edit.handler';
