+++ libs/models/src/lib/employee-recurring-expense.model.ts
@@ -1,0 +1,16 @@
+import { BaseEntityModel as IBaseEntityModel } from './base-entity.model';
+import {
+	RecurringExpenseDeletionEnum,
+	RecurringExpenseModel
+} from './recurring-expense.model';
+
+export interface EmployeeRecurringExpense extends RecurringExpenseModel {
+	employeeId: string;
+}
+
+export interface EmployeeRecurringExpenseByMonthFindInput
+	extends IBaseEntityModel {
+	employeeId?: string;
+	month?: number;
+	year?: number;
+}
