+++ libs/models/src/lib/organization-recurring-expense.model.ts
@@ -1,6 +1,8 @@
 import { BaseEntityModel as IBaseEntityModel } from './base-entity.model';
-import { RecurringExpenseDeletionEnum } from './employee-recurring-expense';
-import { RecurringExpenseModel } from './recurring-expense.model';
+import {
+	RecurringExpenseModel,
+	RecurringExpenseDeletionEnum
+} from './recurring-expense.model';
 
 export interface OrganizationRecurringExpense extends RecurringExpenseModel {
 	orgId: string;
@@ -16,29 +18,9 @@ export interface OrganizationRecurringExpenseFindInput
 	currency?: string;
 }
 
-/**
- * categoryName, value and currency are the new values.
- * startDay, startMonth, startYear represent the date when this was edited.
- */
-export interface OrganizationRecurringExpenseEditInput
-	extends IBaseEntityModel {
-	startDay: number;
-	startMonth: number;
-	startYear: number;
-	categoryName: string;
-	value: number;
-	currency: string;
-}
-
 export interface OrganizationRecurringExpenseByMonthFindInput
 	extends IBaseEntityModel {
 	orgId?: string;
 	month?: number;
 	year?: number;
-}
-
-export interface OrganizationRecurringExpenseDeleteInput {
-	deletionType: RecurringExpenseDeletionEnum;
-	month: number;
-	year: number;
 }
